To remove legacy plugin from code base, leaving in just new impl.
