--- src/main/groovy/com/github/prokod/gradle/crossbuild/BridgingExtension.groovy
@@ -1,30 +1,0 @@
-/*
- * Copyright 2016-2017 the original author or authors
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package com.github.prokod.gradle.crossbuild
-
-import org.gradle.api.Project
-
-/**
- * As it's name suggests, This extension is used to "bridge" between "Project" space and "Model" space,
- *  injecting {@link Project} to the "Model" space
- */
-class BridgingExtension {
-    final Project project
-
-    BridgingExtension(Project project) {
-        this.project = project
-    }
-}
