--- src/main/resources/META-INF/gradle-plugins/com.github.prokod.gradle-crossbuild1.properties
@@ -1,1 +1,0 @@
-implementation-class=com.github.prokod.gradle.crossbuild.CrossBuildPlugin1
